feat: misc improvements and cleanup after hackathon
