+++ libs/repositories/src/index.ts
@@ -40,3 +40,10 @@ export * from './TokenHolderRepository/TokenHolderRepositoryMoralis';
 export * from './SimulationRepository/SimulationRepository';
 export * from './SimulationRepository/SimulationRepositoryTenderly';
 export * from './SimulationRepository/tenderlyTypes';
+
+// Indexer state repository
+export * from './IndexerStateRepository/IndexerStateRepository';
+
+// Notifications repositories
+export * from './PushNotificationsRepository/PushNotificationsRepository';
+export * from './PushSubscriptionsRepository/PushSubscriptionsRepository';
